feat: 文章更新算法优化

+++ sai-web/src/main/java/ink/whi/web/admin/rest/TagSettingsRestController.java
@@ -68,8 +68,8 @@ public class TagSettingsRestController {
     }
 
     @GetMapping(path = "list")
-    public ResVo<PageVo<TagDTO>> list(@RequestParam(name = "pageNumber", required = false) Integer pageNumber,
-                                      @RequestParam(name = "pageSize", required = false) Integer pageSize) {
+    public ResVo<PageVo<TagDTO>> list(@RequestParam(name = "pageNumber", required = false) Long pageNumber,
+                                      @RequestParam(name = "pageSize", required = false) Long pageSize) {
         pageNumber = NumUtil.nullOrZero(pageNumber) ? 1 : pageNumber;
         pageSize = NumUtil.nullOrZero(pageSize) ? 10 : pageSize;
         PageVo<TagDTO> tagDTOPageVo = tagSettingService.getTagList(PageParam.newPageInstance(pageNumber, pageSize));
